[ refactor ] 기존 페이지 폴더 삭제 및 컴포넌트 폴더 내 페이지 구분
