+++ src/components/CollectDiaryPage/NothingPage/index.tsx
@@ -1,0 +1,62 @@
+import styled from 'styled-components';
+import Button from '../../Common/SubmitForm/Button';
+import { useNavigate } from 'react-router-dom';
+
+const NothingPage = () => {
+  const navigate = useNavigate();
+
+  return (
+    <Container>
+      <Title>일기 모아보기</Title>
+      <div>
+        <header>
+          아직 작성된 일기가 없어요.
+          <br />
+          오늘 날짜로 일기를 써보세요.
+        </header>
+        <Button
+          onclick={() => navigate('/main')}
+          placeholder={'일기쓰기'}
+          isValid={true}
+          width={16}
+        />
+      </div>
+    </Container>
+  );
+};
+
+export default NothingPage;
+
+const Title = styled.h1`
+  font-size: 2rem;
+  margin-bottom: 2rem;
+  color: #333333;
+  font-family: Pretendard;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 30px;
+  padding: 2rem 2rem 0 2rem;
+`;
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  gap: 23.6rem;
+  div {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 23.6rem;
+    header {
+      color: #a1a1a1;
+      text-align: center;
+      font-family: Pretendard;
+      font-size: 16px;
+      font-style: normal;
+      font-weight: 400;
+      line-height: 24px;
+    }
+  }
+`;
