+++ src/components/CollectPinPage/index.tsx
@@ -1,0 +1,3 @@
+export default function CollectPinPage() {
+  return <div></div>;
+}
