+++ src/components/Common/NotFoundPage/index.tsx
@@ -1,0 +1,3 @@
+export default function NotFoundPage() {
+  return <div>NotFoundPage</div>;
+}
