--- src/components/Common/index.tsx
@@ -1,5 +1,0 @@
-const index = () => {
-  return <div>index</div>;
-};
-
-export default index;
