--- src/components/Diary/EditDiary.tsx
@@ -1,3 +1,0 @@
-export default function EditDiary() {
-  return <div>EditDiary</div>;
-}
