+++ src/components/DiaryPage/EditDiaryPage/index.tsx
@@ -1,0 +1,3 @@
+export default function EditDiaryPage() {
+  return <div>EditDiary</div>;
+}
