--- src/pages/CollectDiary.tsx
@@ -1,7 +1,0 @@
-import CollectDiaryView from "../components/CollectDiary";
-
-const CollectDiary = () => {
-  return <CollectDiaryView />;
-};
-
-export default CollectDiary;
