--- src/pages/CollectPin.tsx
@@ -1,5 +1,0 @@
-const CollectPin = () => {
-  return <div>CollectPin</div>;
-};
-
-export default CollectPin;
