--- src/pages/Diary.tsx
@@ -1,5 +1,0 @@
-import WriteDiary from "../components/Diary/WriteDiary";
-
-export default function Diary() {
-  return <WriteDiary />;
-}
