--- src/pages/Landing.tsx
@@ -1,11 +1,0 @@
-import LandingView from "../components/Landing/LandingView.tsx";
-
-const Landing = () => {
-  return (
-    <>
-      <LandingView />
-    </>
-  );
-};
-
-export default Landing;
