--- src/pages/Login.tsx
@@ -1,7 +1,0 @@
-import LoginView from "../components/Login/LoginView";
-
-const Login = () => {
-  return <LoginView />;
-};
-
-export default Login;
