--- src/pages/Main.tsx
@@ -1,8 +1,0 @@
-import MainView from '../components/Main/MainView';
-
-const Main = () => {
-  
-  return <MainView />;
-};
-
-export default Main;
