--- src/pages/NotFoundPage.tsx
@@ -1,5 +1,0 @@
-const NotFoundPage = () => {
-  return <div>NotFoundPage</div>;
-};
-
-export default NotFoundPage;
