--- src/pages/SignUp.tsx
@@ -1,7 +1,0 @@
-import SignUpView from "../components/SignUp/SignUpView";
-
-const SignUp = () => {
-  return <SignUpView />;
-};
-
-export default SignUp;
